refactor(schema-compiler): Extract testWithPreAggregation helper function
